requirements.txt
feedgen==0.9.0
lxml==4.9.2
python_dateutil==2.8.2
pytz==2019.3
requests==2.22.0
xmltodict==0.13.0
